[fix] Remove dynamic creation of logger.
This was causing a memory leak, because a lots of logger were created
and they were never freed since this is how the logging framework for
python works.

Also standardize logger usage, i.e. is always named "logger" and is used
as logger.info("%s", 1) instead of logger.info("%s" % 1).