[fix] hande SSL_ERROR_WANT_READ on recv from CTI client.
This can happens because we are using SSL/TLS connection with an underlying
non blocking socket.

In these cases, the socket can sometimes be ready for reading as seen by
select(), but it contains only partial SSL information in it, so an
SSLError is raised when calling recv().

The problems happens on Windows / Mac version of XiVO client when
there's a certain amount (50~) of users in the xivo.