[#4237] Rename test and remove some assertion
The purpose of these tests is to make sure that no exception is raised when
the parse function is called. Asserting that command is not None is miss leading
on the purpose of the test.

Appended _no_exception to the test case names to indicate that we are checking
if the command can be instanciated from_dict without throwing an exception.