[#4257] fix and unit-test channel variable inheritance
Problem was: child channel that inherited from parent channel shared a same dict
instance with its parent, so updating a value in the parent channel also changed
the value in the child channel. In this case, calleridname and calleridnum of
one channel was replaced with callerid{name,num} from another channel.