track call form variables by uniqueid
Remove extra_data from the Channel class
Add a VariableAggregator to the call_forms package
Quickfix the ChannelDAO that rely on call form variables to get the caller ID of a channel
Change sheets to work with unique ids instead of channels

This _BIG_ change is the result of failing to get a working call form flow when
basing calls on the channel name. Local channels and masquerades made it impossible
to follow a simple call to a queue.

Using the uniqueid remove all this mess since a call form has the same unique id
at each step of it's lifetime.

This change also remove the only use of the xivo_manager_inherit patch in asterisk