a Call is now made of two _Channels
a _Channel is a package private class that contains call related
information, the extension and the channel name.

only the extension of a _Channel should be accessed outside the xivo.call
package.

Calls are also tracked earlier, starting at the NewCall event and are
modified while the call is progressing.