merge append_msg and send_message together
Also, only append message to queue if not self.transferconnection. This
way, messages are not sent on connection used to transfer faxes from the
client to the server. It is not useful to send message to these
connections since the client never reads from it (plus, it already
receives the message on its other, non fax connection).