prevent "missed call" display on switchboard retrieve
Before, when a call was retrieved from the switchboard, the phone
(tested on Aastra and Yealink) would display the call as a missed call,
since the channel is first hanged up.

This commit prevent this call from being displayed as a missed call.
This was causing some usability problem with the Yealink.