AsyncRunner: don't run callback if exception occured
Before, if an exception was raised in function, the callback would be
run with a None result. Code using run_with_cb was not written to handle
a None result, so I believe the intent of run_with_cb was that if
function raised an exception, the callback would not be run.