current_call: always update call on local optimization
Previously, we were only updating the calls if we were able to get the 4
calls that were related to the "local channel optimization" event. If
one of the calls were missing (for example, the source_call), we were
not updating the destination call, which was eventually causing
functional problems to the switchboard operator, like not being able to
put its current call on hold.

Calls are sometimes missing from calls_per_line because the current call
manager is not tracking correctly calls when there is multiple channels
(i.e. calls) involved for a given lines.

Ticket #5882.