use explicit buffer size for CTI connection
If a CTI connection receives a TLS record with a very large payload,
xivo-ctid could in theory be in the case where:
- select(2) doesn't report the socket as being "read-ready"
- but there is still data to be received from sock.recv()

And thus, requests from the client would not be processed in a timely
fashion.

So use a per-connection type buffer size, to make this weird constraint
a bit more explicit.

Also, reduce the value of non-CTI non-AMI connection type, which was
unnecessarily high.