disconnection: extract a disconnect_iface
add a method on the ctiserver class to be able to disconnect and remove
a connection (interface_* instance). This avoids the whole return value
shenanigan between the callbacks and the ctiserver class. See
cti_reply_generator.py and cti_command_runner.py to see what I am trying
to avoid/remove.