service_discovery: move all logic to a context manager
the context manager checks if the service is running (not implemented in
this commit) then registers on consul and sends a bus message. When
done, it deregisters the service and publish a deregister message on the
bus.

The check will be used to avoid registering before the service is ready
to handle http requests or cti connections.