reorder operations when changing presence
when the user changes it's presence using the xivo client, the first
step is to send a bus message to notify that its status has
changed. When xivo-ctid receives one such message, sent by itself or by
xivo-ctid-ng, it applies the change and does all operation that are
required based on the configured presence options.

this will allow the presence to be changed by xivo-ctid-ng by publishing
a bus message without having to duplicate the logic of changing a
presence.