async_runner: add _on_response and _on_error params to run
this new interface allow the user to handle errors with a callback
without breaking the current implementation.

the method run_with_cb is not necessary anymore and could be removed
when all usage use the new run params.

Example:

>>> runner.run(foo, 'ab', 42, foo='bar',
               _on_response=on_response_handler,
               _on_error=on_error_handler)

will execute foo('ab', 42, foo='bar') in the thread and schedule a task
on the main thread to handle the response or the error.