current_call: fix bug that mark completed transfer as answered
There are a couple of problems with the transfers events from xivo-bus.

- some of them have the same routing key
- the unmarshall does not check if the message correspond to the current
class

When the event name is not checked and a CompleteTransferEvent is
received (same routing key) the transfer is marked as answered again and
a message is sent to the client which enables the complete button. If
the button is enabled before the Call button is enabled theres no
problem. If the message is received after that change, the switchboard
is left in an unusable state with the Complete and Cancel button enabled
but no transfer to complete.