always declare the collectd bus exchange
reason: when an event is sent to an inexisting exchange, RabbitMQ
sends a Channel.Close method, but kombu Producers never read messages,
and continue sending events, even if the channel is already closed
from RabbitMQ. Note that the kombu Connection is still "alive", because the
underlying TCP socket is still connected, but is unusable for sending,
as long as nobody consumes the Channel.Close method.

The exchange parameters must be the same with collectd, else when trying
to re-declare the exchange with different parameters, RabbitMQ will
reject the declaration.