switchboard stats: fix stats after hold
The new way to hold a channel is to use xivo-ctid-ng to do a blind
transfer to the hold queue.  This changes the old behaviour where a call
would keep the same linked_id for all of its lifetime.

Hold transfer flow goes as follow:

1. When doing the transfer using ctid-ng the transfer_id is sent to the
stats dispatcher which will place this call on hold instead of marking
as transferred when the transfer completed event is received.
2. When the channel variable XIVO_TRANSFER_ROLE=recipient is received,
the dispatcher will keep track of its linked_id.
3. When the channel variable XIVO_TRANSDER_ID is received for a channel
with a receipient role "recipient" it is added to a map of transfer_id
to linked_id.
4. When the transfer completed event is received, if it's id is in the
hold transfer, it will be marked as on_hold. If a new linked_id has been
identified with this transfer_id, the maps are updated to follow this
new linked id in the dispatcher and in the switchboard.