starting handling multi-devices hint
With this change, when the status of a hint with multiple lines
("devices" in asterisk terminology) changes, a "status change" is sent
for all the lines composing the hint.

This is conceptually incorrect, since in most case only 1 line will have
its status changed, but since we don't have yet a notion of a user
"telephony status", and in most of our cases we are using the "line
status" as an indicator of the "telephony status" of the associated
user, we did this.

This should eventually be changed to separate the concept more clearly
(i.e. "line status" vs "user telephony status", which is not the same
thing when a user has more than 1 line).